Fix issue where first row was being cut off
Apparently, the course listing website no longer shows a nav row at the top of the table on every page. (Now it's only at the bottom.) I was not including the first row on every page (starting at page 2) in order to not include the nav row. Since they deleted
the nav row at the top, this had resulted in the first row of the actual results on pages 2+ being deleted.